pyproject.toml
[tool.poetry]
name = "fluidstack"
version = "0.0.0"
description = ""
readme = "README.md"
authors = []
keywords = []

classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
packages = [
    { include = "fluidstack", from = "src"}
]

[project.urls]
Repository = 'https://github.com/fluidstackio/fluidstack-python-sdk'

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.21.2"
pydantic = ">= 1.9.2"
typing_extensions = ">= 4.0.0"

[tool.poetry.dev-dependencies]
mypy = "1.9.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.23.5"
python-dateutil = "^2.9.0"
python-dotenv = "^1.0.1"


[tool.pytest.ini_options]
testpaths = [ "tests" ]
asyncio_mode = "auto"

[tool.mypy]
plugins = ["pydantic.mypy"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
